file moves and links to editor :
