+++ toolbar-extender/sitefinity-videos-toolbar-item-provider.css
